server/lambda/requirements.txt
requests
python-docx
lxml